Save the position of the last watch replacement (#42) and more
- merge assign.rs and var.rs into a module
- better modularity and abstraction via traits
- add and refactor restart module: `GeometricStabilizer`, `ProgressLBD`
- change a lot of options' meanings and mnemonics
- add some features for development: boundary_check, use_core, EVSIDS, trace_analysis
- `propagate` skips up to the last position
- remove code about 'deep search' completely, superseded with stabilizing mode
- stop *too* dynamic var decay control
- change the activation timings of simplification
- stats data are stored in each module
- activate `JUST_USED` for clause reduction
