+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "splr"
-version = "0.3.3-dev.0"
+version = "0.4.0-RC1"
 authors = ["Narazaki Shuji <shujinarazaki@protonmail.com>"]
 description = "A modern CDCL SAT solver in Rust"
 edition = "2018"
@@ -18,7 +18,7 @@ libc = "0.2"
 structopt = "0.3"
 
 [features]
-default = ["use_core", "EVSIDS"]
+default = []
 EVSIDS = []
 ema_calibration = []
 boundary_check = []
