+++ src/assign/select.rs
@@ -1,0 +1,48 @@
+/// Var selection
+use super::{AssignStack, VarHeapIF, VarOrderIF};
+#[cfg(feature = "use_core")]
+use crate::state::SearchStrategy;
+/// Crate `propagator` implements Boolean Constraint Propagation and decision var selection.
+/// This version can handle Chronological and Non Chronological Backtrack.
+use crate::types::*;
+
+/// ```
+/// let x: Lbool = var_assign!(self, lit.vi());
+/// ```
+macro_rules! var_assign {
+    ($asg: expr, $var: expr) => {
+        unsafe { *$asg.assign.get_unchecked($var) }
+    };
+}
+
+/// API for var selection.
+pub trait VarSelectIF {
+    /// select a new decision variable.
+    fn select_var(&mut self) -> VarId;
+    /// update the internal heap on var order.
+    fn update_order(&mut self, v: VarId);
+    /// rebuild the internal var_order
+    fn rebuild_order(&mut self);
+}
+
+impl VarSelectIF for AssignStack {
+    fn select_var(&mut self) -> VarId {
+        loop {
+            let vi = self.get_heap_root();
+            if var_assign!(self, vi).is_none() && !self.var[vi].is(Flag::ELIMINATED) {
+                return vi;
+            }
+        }
+    }
+    fn update_order(&mut self, v: VarId) {
+        self.update_heap(v);
+    }
+    fn rebuild_order(&mut self) {
+        self.var_order.clear();
+        for vi in 1..self.var.len() {
+            if var_assign!(self, vi).is_none() && !self.var[vi].is(Flag::ELIMINATED) {
+                self.insert_heap(vi);
+            }
+        }
+    }
+}
